Put CutSelector in its own component
